bunch of small fixes, see the pull request notes
