+++ README.md
@@ -1,2 +1,7 @@
 # UVStocks
-A simple stock trading game.
+A simple stock trading game. Shoot for a high score!
+
+## Running the program
+
+Just clone the repository and run uvstocks.py! 
+
