+++ requirements.txt
@@ -1,0 +1,2 @@
+matplotlib>=3.5.1
+Pillow>=9.2.0
